refactor: moved validation-service to auth-module
