+++ src/app.module.ts
@@ -10,7 +10,6 @@ import { LoggerMiddleware } from './logger/logger.middleware';
 import { CalendarModule } from './calendar/calendar.module';
 import { AstraModule } from '@cahllagerfeld/nestjs-astra';
 import { AstraConfigService } from './astra/astra-config.service';
-import { CommonsModule } from './commons/commons.module';
 
 @Module({
   imports: [
@@ -25,7 +24,6 @@ import { CommonsModule } from './commons/commons.module';
     }),
     AuthModule,
     CalendarModule,
-    CommonsModule,
   ],
   controllers: [AppController],
   providers: [AppService],
