+++ src/calendar/calendar.module.ts
@@ -1,12 +1,12 @@
 import { AstraModule } from '@cahllagerfeld/nestjs-astra';
 import { Module } from '@nestjs/common';
-import { CommonsModule } from '../commons/commons.module';
+import { AuthModule } from '../auth/auth.module';
 import { CalendarController } from './calendar.controller';
 import { CalendarService } from './calendar.service';
 
 @Module({
   imports: [
-    CommonsModule,
+    AuthModule,
     AstraModule.forFeature({ namespace: 'eddiehub', collection: 'calendar' }),
   ],
   controllers: [CalendarController],
