--- src/commons/commons.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { ValidationService } from './validation.service';
-
-@Module({
-  providers: [ValidationService],
-  exports: [ValidationService],
-})
-export class CommonsModule {}
