--- src/commons/validation.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { ValidationService } from './validation.service';
-
-describe('ValidationService', () => {
-  let service: ValidationService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [ValidationService],
-    }).compile();
-
-    service = module.get<ValidationService>(ValidationService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
