+++ src/discord/discord.module.ts
@@ -1,13 +1,11 @@
 import { Module } from '@nestjs/common';
 import { DiscordService } from './discord.service';
 import { DiscordController } from './discord.controller';
-import { AuthModule } from '../auth/auth.module';
 import { AstraModule } from '@cahllagerfeld/nestjs-astra';
-import { CommonsModule } from '../commons/commons.module';
+import { AuthModule } from '../auth/auth.module';
 
 @Module({
   imports: [
-    CommonsModule,
     AuthModule,
     AstraModule.forFeature({ namespace: 'eddiehub', collection: 'discord' }),
   ],
