+++ src/discord/discord.service.spec.ts
@@ -2,7 +2,7 @@ import { AstraModule } from '@cahllagerfeld/nestjs-astra';
 import { ConfigModule } from '@nestjs/config';
 import { Test, TestingModule } from '@nestjs/testing';
 import { AstraConfigService } from '../astra/astra-config.service';
-import { CommonsModule } from '../commons/commons.module';
+import { AuthModule } from '../auth/auth.module';
 import { DiscordService } from './discord.service';
 
 describe('DiscordService', () => {
@@ -11,7 +11,7 @@ describe('DiscordService', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       imports: [
-        CommonsModule,
+        AuthModule,
         ConfigModule.forRoot({
           isGlobal: true,
         }),
