+++ src/github/github.service.spec.ts
@@ -3,6 +3,7 @@ import { HttpModule } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { Test, TestingModule } from '@nestjs/testing';
 import { AstraConfigService } from '../astra/astra-config.service';
+import { AuthModule } from '../auth/auth.module';
 import { CommunitystatsMappingService } from './communitystats-mapping.service';
 import { GeocodingService } from './geocoding.service';
 import { GithubService } from './github.service';
@@ -13,6 +14,7 @@ describe('GithubService', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       imports: [
+        AuthModule,
         HttpModule,
         ConfigModule.forRoot({
           isGlobal: true,
