+++ src/standup/standup.controller.spec.ts
@@ -3,7 +3,6 @@ import { ConfigModule } from '@nestjs/config';
 import { Test, TestingModule } from '@nestjs/testing';
 import { AstraConfigService } from '../astra/astra-config.service';
 import { AuthModule } from '../auth/auth.module';
-import { CommonsModule } from '../commons/commons.module';
 import { StandupController } from './standup.controller';
 import { StandupService } from './standup.service';
 
@@ -13,7 +12,6 @@ describe('StandupController', () => {
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
       imports: [
-        CommonsModule,
         AuthModule,
         ConfigModule.forRoot({
           isGlobal: true,
