+++ src/standup/standup.module.ts
@@ -3,11 +3,9 @@ import { AuthModule } from '../auth/auth.module';
 import { StandupController } from './standup.controller';
 import { StandupService } from './standup.service';
 import { AstraModule } from '@cahllagerfeld/nestjs-astra';
-import { CommonsModule } from '../commons/commons.module';
 
 @Module({
   imports: [
-    CommonsModule,
     AuthModule,
     AstraModule.forFeature({ namespace: 'eddiehub', collection: 'standup' }),
   ],
